changed cuisine constants and ingredient amount to float

+++ backend/app.py
@@ -260,7 +260,7 @@ def get_ingredients_for_specific_recipe(recipe_id: int):
                     "id": <integer>,
                     "recipe_id": <integer>,
                     "name": <string>,
-                    "amount": <integer>,
+                    "amount": <float>,
                     "unit": <string>
                 },
                 ...
@@ -638,7 +638,7 @@ def post_new_recipe_for_user(user_id: int):
             "ingredients": [
                 {
                     "name": <string>,
-                    "amount": <integer>,
+                    "amount": <float>,
                     "unit": <string>
                 },
                 ...
@@ -768,7 +768,7 @@ def update_recipe(recipe_id):
             "ingredients": [
                 {
                     "name": <string>,
-                    "amount": <integer>,
+                    "amount": <float>,
                     "unit": <string>
                 },
                 ...
